web_service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pyyaml>=6.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
termcolor>=2.0.0
PyQt6>=6.6.0
tokenizers>=0.15.0
natsort>=8.1.0
onnx>=1.15.0
onnxruntime>=1.17.0
scipy>=1.4.1
shapely>=2.0.0
lapx>=0.5.5
json_repair>=0.29.0
jsonlines>=4.0
matplotlib>=3.0.0
importlib_metadata>=7.0
requests>=2.28.0
